Move majority of input.c to UI
