+++ src/openrct2/Context.h
@@ -223,6 +223,7 @@ extern "C"
     void context_broadcast_intent(Intent * intent);
     void context_force_close_window_by_class(rct_windowclass wc);
     void context_update_map_tooltip();
+    void context_handle_input();
     void context_input_handle_keyboard(bool isTitle);
     bool context_read_bmp(void * * outPixels, uint32 * outWidth, uint32 * outHeight, const utf8 * path);
     void context_quit();
