+++ src/openrct2/ui/WindowManager.h
@@ -43,6 +43,7 @@ namespace OpenRCT2
             virtual rct_window * ShowError(rct_string_id title, rct_string_id message) abstract;
             virtual void ForceClose(rct_windowclass windowClass) abstract;
             virtual void UpdateMapTooltip() abstract;
+            virtual void HandleInput() abstract;
             virtual void HandleKeyboard(bool isTitle) abstract;
             virtual std::string GetKeyboardShortcutString(sint32 shortcut) abstract;
         };
